新建统计数据表，定时收集统计信息

+++ clever-task/src/main/java/org/clever/task/manage/model/response/StatisticsInfoRes.java
@@ -17,7 +17,7 @@ public class StatisticsInfoRes {
     /**
      * 触发次数
      */
-    private Integer triggerCount;
+    private Long triggerCount;
     /**
      * 命名空间数
      */
